# api-documentation-old.yaml
openapi: 3.0.0
info:
  title: InCrowd API
  version: 1.0.0
  description: |
    # InCrowd API - Documentazione Completa
    
    Benvenuto nella documentazione dell'API di InCrowd, una piattaforma innovativa di crowdsourcing per proposte e iniziative cittadine.
    
    ## Panoramica
    
    L'API InCrowd fornisce accesso completo alla piattaforma per:
    - Gestione utenti (cittadini privati ed enti)
    - Creazione e gestione proposte
    - Sistema di follow/followers
    - Amministrazione e moderazione
    
    ## Autenticazione e Sicurezza
    
    L'API utilizza autenticazione JWT (JSON Web Token):
    
    1. **Login**: Ottieni il token tramite `POST /api/auth/login`
    2. **Utilizzo**: Includi il token nell'header di ogni richiesta:
       ```
       Authorization: Bearer <your-jwt-token>
       ```
    3. **Scadenza**: I token hanno durata limitata, effettua refresh quando necessario
    
    ## Struttura delle Response
    
    Tutte le risposte seguono il formato standardizzato:
    
    ```json
    {
      "data": { /* Dati della risposta */ },
      "message": "Messaggio descrittivo",
      "error": null // Solo presente in caso di errori
    }
    ```
    
    ## Gestione Errori
    
    - **400 Bad Request**: Dati non validi o parametri mancanti
    - **401 Unauthorized**: Token mancante, scaduto o non valido
    - **403 Forbidden**: Permessi insufficienti per l'operazione
    - **404 Not Found**: Risorsa non trovata
...
    
    ## Tipi di Utente
    
...
    
...
    
...

...
servers:
...
paths:
...
  /api/auth/password:
    patch:
      tags: [Authentication]
      operationId: updatePassword
      summary: Aggiorna password utente
      description: |
        Modifica la password dell'utente autenticato.
        Disponibile per tutti i tipi di utente (privato, ente, operatore).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - newPassword
              properties:
                newPassword:
                  type: string
                  minLength: 8
                  pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$"
                  description: "Nuova password (minimo 8 caratteri, 1 maiuscola, 1 minuscola, 1 numero e 1 carattere speciale)"
                  example: "NewSecurePass123!"
                currentPassword:
                  type: string
                  description: "Password attuale (richiesta per sicurezza)"
                  example: "OldSecurePass123!"
      responses:
        '200':
          description: Password aggiornata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                message: "Password updated successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # =============================================================================
  # USER ENDPOINTS (UNIFICATI: privati + enti)
  # =============================================================================
  /api/user:
    get:
      tags: [Users]
      operationId: getAllUsers
      summary: Recupera tutti gli utenti (solo operatori)
      description: |
        Endpoint riservato a operatori e amministratori per visualizzare tutti gli utenti.
        Include sia utenti privati che enti.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: user_type
          schema:
            type: string
            enum: [privato, ente]
          description: "Filtra per tipo di utente"
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Lista degli utenti recuperata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                data:
                  users:
                    - _id: "60d0fe4f5311236168a109ca"
                      nome: "Mario"
                      cognome: "Rossi"
                      user_type: "privato"
                      email: "mario.rossi@example.com"
                      followersCount: 150
                      followingCount: 75
                    - _id: "60d0fe4f5311236168a109cb"
                      nome: "Comune di Trento"
                      user_type: "ente"
                      email: "info@comune.trento.it"
                      codiceFiscale: "12345678901"
                  pagination:
                    currentPage: 1
                    totalPages: 5
                    totalItems: 97
                    itemsPerPage: 20
                message: "Users retrieved successfully"
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Users]
      operationId: createUser
      summary: Registra nuovo utente (privato o ente)
      description: |
        Crea un nuovo account utente. Il tipo viene determinato dai campi forniti:
        - Se presente `cognome` → utente privato
        - Se presente solo `nome` senza `cognome` → ente
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - nome
                - email
                - password
              properties:
                nome:
                  type: string
                  description: "Nome utente o nome ente"
                  example: "Mario"
                cognome:
                  type: string
                  description: "Cognome (solo per utenti privati)"
                  example: "Rossi"
                codiceFiscale:
                  type: string
                  description: "Codice fiscale (privato) o Partita IVA (ente)"
                  example: "RSSMRA80A01H501U"
                email:
                  type: string
                  format: email
                  example: "mario.rossi@example.com"
                password:
                  type: string
                  minLength: 8
                  pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$"
                  description: "Password (minimo 8 caratteri, 1 maiuscola, 1 minuscola, 1 numero e 1 carattere speciale)"
                  example: "SecurePass123!"
                biografia:
                  type: string
                  maxLength: 500
                  description: "Biografia o descrizione"
                  example: "Appassionato di tecnologia e sport"
                fotoProfilo:
                  type: string
                  format: binary
                  description: "Foto profilo (opzionale)"
                oauthCode:
                  type: string
                  description: "Codice Google OAuth (per registrazione Google)"
      responses:
        '201':
          description: Utente creato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                data:
                  user:
                    _id: "60d0fe4f5311236168a109ca"
                    nome: "Mario"
                    cognome: "Rossi"
                    user_type: "privato"
                    email: "mario.rossi@example.com"
                    codiceFiscale: "RSSMRA80A01H501U"
                    biografia: "Appassionato di tecnologia e sport"
                    followersCount: 0
                    followingCount: 0
                message: "User created successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/user/me:
    get:
      tags: [Users]
      operationId: getCurrentUser
      summary: Recupera dati utente corrente
      description: Ottiene i dati dell'utente attualmente autenticato
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Dati utente corrente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              examples:
                privato:
                  summary: "Utente privato"
                  value:
                    data:
                      _id: "60d0fe4f5311236168a109ca"
                      nome: "Mario"
                      cognome: "Rossi"
                      user_type: "privato"
                      email: "mario.rossi@example.com"
                      codiceFiscale: "RSSMRA80A01H501U"
                      biografia: "Appassionato di tecnologia e sport"
                      fotoProfilo:
                        data: "/9j/4AAQSkZJRgABAQAAAQABAAD..."
                        contentType: "image/jpeg"
                      followersCount: 150
                      followingCount: 75
                      createdAt: "2024-01-15T10:30:00Z"
                      updatedAt: "2024-08-10T15:45:00Z"
                    message: "User profile retrieved successfully"
                ente:
                  summary: "Utente ente"
                  value:
                    data:
                      _id: "60d0fe4f5311236168a109cb"
                      nome: "Comune di Trento"
                      user_type: "ente"
                      email: "info@comune.trento.it"
                      codiceFiscale: "12345678901"
                      biografia: "Amministrazione comunale della città di Trento"
                      fotoProfilo:
                        data: "/9j/4AAQSkZJRgABAQAAAQABAAD..."
                        contentType: "image/png"
                      followersCount: 2500
                      followingCount: 0
                      createdAt: "2024-01-10T09:00:00Z"
                      updatedAt: "2024-08-15T12:30:00Z"
                    message: "Organization profile retrieved successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Users]
      operationId: deleteCurrentUser
      summary: Elimina account utente corrente
      description: |
        Elimina definitivamente l'account dell'utente corrente e tutti i dati associati.
        Include proposte e commenti. Non disponibile per operatori e amministratori.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Account eliminato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                message: "Account deleted successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/user/profile:
    patch:
      tags: [Users]
      operationId: updateProfile
      summary: Aggiorna profilo utente corrente
      description: |
        Modifica i dati del profilo dell'utente autenticato.
        I campi disponibili dipendono dal user_type.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - nome
              properties:
                nome:
                  type: string
                  description: "Nome utente o nome ente"
                  example: "Mario"
                cognome:
                  type: string
                  description: "Cognome (solo per utenti privati)"
                  example: "Rossi"
                codiceFiscale:
                  type: string
                  description: "Codice fiscale (privato) o Partita IVA (ente)"
                  example: "RSSMRA80A01H501U"
                biografia:
                  type: string
                  maxLength: 500
                  description: "Biografia o descrizione"
                  example: "Appassionato di tecnologia e sport"
                fotoProfilo:
                  type: string
                  format: binary
                  description: "Nuova foto profilo (opzionale)"
      responses:
        '200':
          description: Profilo aggiornato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                data:
                  user:
                    _id: "60d0fe4f5311236168a109ca"
                    nome: "Mario"
                    cognome: "Rossi"
                    user_type: "privato"
                    email: "mario.rossi@example.com"
                    codiceFiscale: "RSSMRA80A01H501U"
                    biografia: "Appassionato di tecnologia e sport aggiornato"
                    updatedAt: "2024-08-18T10:30:00Z"
                message: "Profile updated successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/user/search:
    get:
      tags: [Users]
      operationId: searchUsers
      summary: Ricerca utenti
      description: |
        Ricerca utenti per nome, cognome o biografia con paginazione.
        Include sia utenti privati che enti nei risultati.
      parameters:
        - in: query
          name: query
          required: true
          schema:
            type: string
            example: "Mario"
          description: "Termine di ricerca"
        - in: query
          name: user_type
          schema:
            type: string
            enum: [privato, ente]
          description: "Filtra per tipo di utente"
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: "Numero pagina"
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
          description: "Elementi per pagina"
      responses:
        '200':
          description: Risultati ricerca utenti
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                data:
                  users:
                    - _id: "60d0fe4f5311236168a109ca"
                      nome: "Mario"
                      cognome: "Rossi"
                      user_type: "privato"
                      biografia: "Appassionato di tecnologia e sport"
                      fotoProfilo:
                        data: "/9j/4AAQSkZJRgABAQAAAQABAAD..."
                        contentType: "image/jpeg"
                      followersCount: 150
                      followingCount: 75
                    - _id: "60d0fe4f5311236168a109cb"
                      nome: "Comune di Trento"
                      user_type: "ente"
                      biografia: "Amministrazione comunale della città di Trento"
                      followersCount: 2500
                      followingCount: 0
                  pagination:
                    currentPage: 1
                    totalPages: 3
                    totalItems: 27
                    itemsPerPage: 10
                message: "Search completed successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/user/{id}:
    get:
      tags: [Users]
      operationId: getUserById
      summary: Recupera utente per ID
      description: |
        Ottiene i dati di un utente specifico. 
        I dati restituiti variano in base all'autenticazione:
        - Non autenticato: solo dati pubblici
        - Operatore: dati completi
        Supporta sia utenti privati che enti.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: "60d0fe4f5311236168a109ca"
          description: "ID dell'utente"
      security:
        - BearerAuth: []
        - {}
      responses:
        '200':
          description: Dati utente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              examples:
                privato:
                  summary: "Utente privato"
                  value:
                    data:
                      _id: "60d0fe4f5311236168a109ca"
                      nome: "Mario"
                      cognome: "Rossi"
                      user_type: "privato"
                      biografia: "Appassionato di tecnologia e sport"
                      fotoProfilo:
                        data: "/9j/4AAQSkZJRgABAQAAAQABAAD..."
                        contentType: "image/jpeg"
                      followersCount: 150
                      followingCount: 75
                      createdAt: "2024-01-15T10:30:00Z"
                    message: "User profile retrieved successfully"
                ente:
                  summary: "Utente ente"
                  value:
                    data:
                      _id: "60d0fe4f5311236168a109cb"
                      nome: "Comune di Trento"
                      user_type: "ente"
                      biografia: "Amministrazione comunale della città di Trento"
                      followersCount: 2500
                      followingCount: 0
                      createdAt: "2024-01-10T09:00:00Z"
                    message: "Organization profile retrieved successfully"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/user/{id}/avatar:
    get:
      tags: [Users]
      operationId: getUserAvatar
      summary: Recupera avatar utente
      description: Ottiene l'avatar di un utente specifico in formato base64
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: "60d0fe4f5311236168a109ca"
          description: "ID dell'utente"
      responses:
        '200':
          description: Avatar utente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                data:
                  fotoProfilo:
                    data: "/9j/4AAQSkZJRgABAQAAAQABAAD..."
                    contentType: "image/jpeg"
                message: "Avatar retrieved successfully"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

# Rest of the endpoints will continue in the next part...
# I'll continue with Follow, Proposte, Operatori, Admin sections

  # =============================================================================
  # FOLLOW ENDPOINTS
  # =============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token ottenuto tramite login"
...
  schemas:
...
    StandardResponse:
      type: object
      properties:
        data:
          type: object
          description: "Contenuto della risposta (opzionale in caso di errore)"
        message:
          type: string
          description: "Messaggio informativo sempre presente"
          example: "Operation completed successfully"
        error:
          type: object
          description: "Dettagli errore (solo in caso di errore)"
          properties:
            code:
              type: string
              example: "ERROR_CODE"
            details:
              type: string
              example: "Detailed error description"
      required:
        - message

    # =============================================================================
    # USER SCHEMA (UNIFICATO)
    # =============================================================================
...
security:
  - BearerAuth: []
  - {}